Generate the assets for the hydra article
One of the benefits of JS is that I can run the exact
same code server-side to produce images that match the
client-side ones. One of the drawbacks is that I hate
JS.

Took quite a while to figure out how to make code
compatible with browser JS and node.js, but eventually
got it working, and pulled out that business into
hydra_lib.js. The small remainder of leftover code
is in js/hydra_main.js.

Using that lib, I wrote a node.js file that would
generate assets that look identical to what shows up
client-side in the interactive region. The new diagrams
look pretty similar, but of course, the SVG content is
wildly different, so the git diff isn't useful here.

The only asset I haven't generated yet is the last one,
with the ordinals. If I can figure out how to embed
MathJAX content into these SVGs I shall become like unto
a god.