Added summary markers back to posts
There's a plugin for this, but unfortunately, it seems to be broken. I'm
not yet familiar enough with Pelican to understand why this works, or if
it's the proper fix, but if I set the metadata instead of `_summary` in
the plugin's `extract_summary` method, it works.

Also, the summary doesn't respect the MathJaX, since that script is
added at the end. So for now, I'm putting the script in the bottom of
all index-like pages. This should be fine; I don't think there's any of
those pages without math. But it's still gross :(