Add hydra to builder
The new lib/ folder is for code that will be used both for asset
generation, and on the site itself. Haven't gotten the latter working
yet, but we know what's coming (because I wrote a prototype earlier and
this stack is just rehashing it into a sane history...).

Also, the initialization of mathjax and svg.js isn't re-entrant, so I
had to comment out the dehn import. That's next up on the list, I guess.