Allow both builder files to run
The trick here is that we need to clear the canvas. I don't know why
calling shrinkCanvas twice in a row breaks things, but whatever. That's
a problem we can solve later.

I manually verified that the only change in hydra/example-1.svg was the
renaming of SVG ids, and the website still looks normal.